Set up test packages and reorganized tree
