+++ src/components/Card.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import emoji from 'emoji-dictionary';
 
-import './Card.css';
+import './style/Card.css';
 
 class Card extends Component {
   render() {
