+++ src/components/style/Board.css
